Refactor: bunch of cleanup
